backend/requirements.txt
fastapi==0.135.3
uvicorn[standard]
google_api_python_client==2.183.0
langchain_groq==1.1.2
langgraph==1.1.4
protobuf==7.34.1
pydantic==2.12.5
python-dotenv==1.2.2
python_bcrypt==0.3.2
python_jose==3.5.0
qdrant_client==1.17.1
Requests==2.33.1
sentence_transformers==5.3.0
SQLAlchemy==2.0.48
psycopg2-binary
PyMuPDF==1.23.7
psycopg2-binary
python-multipart